src/experimentation.py
```python
# Importing necessary libraries
from google.cloud import aiplatform
from dvc.repo import Repo
from h2oai.client import Client
from mlflow import log_metric, log_param, log_artifacts

class Experimentation:
    def __init__(self, model):
        self.model = model

    def google_ai_platform_experiment(self, project_id, region, experiment_id):
        """
        Function to run an experiment using Google Cloud AI Platform
        """
        client = aiplatform.gapic.JobServiceClient(client_options={"api_endpoint": f"{region}-aiplatform.googleapis.com"})
        parent = f"projects/{project_id}/locations/{region}"
        experiment = client.get_hyperparameter_tuning_job(name=f"{parent}/hyperparameterTuningJobs/{experiment_id}")
        print(experiment)

    def dvc_experiment(self, data_file, params_file):
        """
        Function to run an experiment using DVC
        """
        repo = Repo('.')
        pipeline = repo.create_pipeline(stage_name='train', cmd='python train.py', deps=[data_file, params_file])
        repo.reproduce('train')
        metrics = repo.metrics.show()
        print(metrics)

    def h2oai_experiment(self, server_url, username, password, dataset):
        """
        Function to run an experiment using H2O.ai Driverless AI
        """
        client = Client(server_url, username, password)
        experiment = client.start_experiment_sync(dataset_key=dataset, target_col='label', is_classification=True)
        print(experiment)

    def mlflow_track(self, param, metric, artifact):
        """
        Function to track experiment parameters, metrics, and artifacts using MLflow
        """
        log_param("param", param)
        log_metric("metric", metric)
        log_artifacts(artifact, "experiment_artifacts")
```
