tests/test_experimentation.py
```python
# Importing necessary libraries
import unittest
from unittest.mock import patch, MagicMock
from src.experimentation import Experimentation

class TestExperimentation(unittest.TestCase):
    def setUp(self):
        self.model = MagicMock()
        self.experimentation = Experimentation(self.model)

    @patch('src.experimentation.aiplatform.gapic.JobServiceClient')
    def test_google_ai_platform_experiment(self, mock_client):
        mock_client.return_value.get_hyperparameter_tuning_job.return_value = 'experiment'
        self.experimentation.google_ai_platform_experiment('project_id', 'region', 'experiment_id')
        mock_client.assert_called_once_with(client_options={"api_endpoint": "region-aiplatform.googleapis.com"})
        mock_client.return_value.get_hyperparameter_tuning_job.assert_called_once_with(name="projects/project_id/locations/region/hyperparameterTuningJobs/experiment_id")

    @patch('src.experimentation.Repo')
    def test_dvc_experiment(self, mock_repo):
        mock_repo.return_value.create_pipeline.return_value = 'pipeline'
        mock_repo.return_value.reproduce.return_value = 'reproduce'
        mock_repo.return_value.metrics.show.return_value = 'metrics'
        self.experimentation.dvc_experiment('data_file', 'params_file')
        mock_repo.assert_called_once_with('.')
        mock_repo.return_value.create_pipeline.assert_called_once_with(stage_name='train', cmd='python train.py', deps=['data_file', 'params_file'])
        mock_repo.return_value.reproduce.assert_called_once_with('train')
        mock_repo.return_value.metrics.show.assert_called_once()

    @patch('src.experimentation.Client')
    def test_h2oai_experiment(self, mock_client):
        mock_client.return_value.start_experiment_sync.return_value = 'experiment'
        self.experimentation.h2oai_experiment('server_url', 'username', 'password', 'dataset')
        mock_client.assert_called_once_with('server_url', 'username', 'password')
        mock_client.return_value.start_experiment_sync.assert_called_once_with(dataset_key='dataset', target_col='label', is_classification=True)

    @patch('src.experimentation.log_param')
    @patch('src.experimentation.log_metric')
    @patch('src.experimentation.log_artifacts')
    def test_mlflow_track(self, mock_log_artifacts, mock_log_metric, mock_log_param):
        self.experimentation.mlflow_track('param', 'metric', 'artifact')
        mock_log_param.assert_called_once_with("param", 'param')
        mock_log_metric.assert_called_once_with("metric", 'metric')
        mock_log_artifacts.assert_called_once_with('artifact', "experiment_artifacts")

if __name__ == '__main__':
    unittest.main()
```
